qmlls.ini generation: check parent build dir before writing it
When we see a qt_add_qml_module(URI org.kde.foo) call, we assume during the
.qmlls.ini generation that the QML module will end up at <build>/org/kde/foo/.
When this assumption doesn't hold, for example when CMakeLists.txt is in
/some/path/to/source/CMakeLists.txt, we added /some/ as build directory
because of a missing check that "path", "to" and "source" are equal to "org",
"kde" and "foo".

Add a check and use the build directory as is when a mismatch between
uri and CML.txt is detected.

Fixes: QTBUG-138599
Change-Id: Ifba0086e70ca9a8f5c9212dd97fb5c8636875090
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 8ba50d5f31772e1e696a1db51585a67b848525d8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>