Propagate attached properties for items within the popup window
The attached property doesn't propagate to the sub-menu items
within the pop when the pop-up type is configured as Popup.Window,
as the attached property, is provided only for the immediate window
(made as part of patch d824585075b0abaec12c0cbb6445aa4eb84b2cea)
and not further up.

This patch ensures that the attached property is provided for the
window if configured for anything further up in the hierarchy.

Fixes: QTBUG-139603
Pick-to: 6.9 6.8
Change-Id: I95b9f735b7e19ad9916562bf09445fd20f2ba1f4
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit e51223c77ca9ff05e7a2db8e7598c762a9918b9b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>