CMake: Work around looking for Qt packages in in-tree tests
Calling find_package(Qt6 COMPONENTS Qml) inside a tests/
CMakeLists.txt with an in-tree tests build will fail while trying to
include the Qt6QmlToolsTargets.cmake file, because the file is
generated at generation time, but we try to include at configure time.

This works in 6.11+, because we have code that tries to detect not
having to include the Targets file if not needed.

Work around the issue by setting QT_NO_CREATE_TARGETS to ON before the
find_package call in the qmltc doc snippet.

Problematic code originally introduced in
5296c0255b83b8a92e82b3ec4d536655742b90cf in dev.
Amends cherry-pick dad91592e609364231decaba771418d2d48d4910

Change-Id: I5ff510c58e9c50cc3d9e8cde999a8ab531debdd5
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>