Fusion: give TextArea a background
This is the same background as TextField. This will cause the
implicitWidth for TextArea to be the same as TextField - i.e. a bit
wider - whereas before it was only as wide as the text + padding.

This fixes it not being visible, especially without placeholder text.

Fixes: QTBUG-139715
Pick-to: 6.9
Change-Id: Ia736c2e282497f0a40bf8726adcc2838082ea5b4
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 006c4eedda9c42b4f47a0b9c5c11277dc341369b)