QML: Preserve return type annotation for function expressions
The parser for QML allows the specification of type hints on a series of
callable elements; such as functions defined as part of a QML type.

It is possible, for example, to provide an hint for the return type of a
function expression, anonymous or not, such as:

```
import QtQml

QtObject {
    property var: function (lhs: int, rhs: int)): int { return lhs + rhs; }
}
```

While this parses correctly, the return type annotation is currently
discarded as part of the parsing process, such that in turn it is
invisible to the engine when managing the function.

Hence, ensure that the return type annotation for function expressions
is preserved during the parsing process so that it can later be visible
to the engine.

In particular, the AST node that represents this kind of element can
already store a type annotation that refers to its return type and later
down the line, the code generation phase is already able to take into
account the annotation when it builds its representation of the function.

Nonetheless, that storage is currently purposefully not used in the
cases we are interested into, albeit it is in other cases that might be
considered more common.
This is most probably, albeit nowhere near certainly, due to simple
historical reasons and partial implementations related to type hints.

Thus, ensure that the return type annotation is correctly preserved and
stored for `FunctionExpression` rules during the parsing process, such
that it can later be recognized and used by the engine.

A few test cases related to the return type annotation were added.

Task-number: QTBUG-137944
Change-Id: I6133bb286a916d0687ff5d5542b9aa769cfd493b
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit cf572e2b0303d681beb00596135f3331e7f03718)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>