Allow text edit to process key events when shortcutoverride configured
The shortcut override has been made to be ignored by default as part of
the patch 1aefea26e5a574dad25646d330a6b3bb943a596c. And the further
changes in the same patch were made to check whether the shortcut
override can be accepted by text edit before processing the events.

If the shortcut override has not been accepted (either by default or
explicitly ignored by the user), the keys would be processed further by
the textedit. The validation added as part of the same patch for
processing the event doesn't seem to be correct, as it allows the key to
be processed by the text edit control only when the key handler is not
configured. This approach seems to be incorrect, as the text edit
control shall be allowed to process the key events regardless of key
handler configuration.

This patch removes that validation, thus allowing the text edit control
to process the key events regardless of key handler configuration.

Fixes: QTBUG-139679
Task-number: QTBUG-136959
Pick-to: 6.8 6.5
Change-Id: I7c066e5f3709a00d0b18f0c5e8b9d1f7944c4e4e
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 30b4419e112643775e788e51fd460fe267093707)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>