Give ownership to qml instead of deleting the object
after calling grabToImage in the callback it returns a qml wrapper
around QQuickItemGrabResult object but then in the Event_Grab_Completed
event it destroys the QQuickItemGrabResult object so the wrapper is
also automatically destroyed after that.
so in order for qml to be able to store the wrapper object and hold the
result as long as needed we give the ownership of QQuickItemGrabResult
to qml instead of deleting it.

Pick-to: 6.8
Fixes: QTBUG-128483
Change-Id: Ibf9b9765e22ee8e0caf8b3646d86f8a4cf789ff1
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 8f0d9f613360ea19af701916a406fd666d07555c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>