qmllint: Also lint inner functions
We were only collecting the 'QML' functions that would be passed to the
AOT compiler. Other function types were ignored.

Also collect those and pass them to the compiler but only use them when
linting. Defer invistigating whether it is a good idea to try to
compile JS functions as well to a later point.

Logic in a few places was adapted with this changing assumption.

[ChangeLog][QML][qmllint] qmllint will now lint inner functions,
defined in javascript, in addition to top-level functions and bindings.

Fixes: QTBUG-138845
Pick-to: 6.8 6.5
Change-Id: If6f62aeace8739442b6a1f355fad95ce19c0643c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit a0a2a239dc8e1a486ae8bdf14ec7fd5db27311e1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>