QtQml: Fix initialization of non-optional static plugins
Don't let the initialization of a non-optional, supposedly dynamic,
plugin pass if there actually is no QPluginLoader for it. The absence of
the QPluginLoader in this case tells us that the plugin is in fact
static and we need to check the static plugin map.

Optional static plugins are not an issue here since there is no code
path to create instances of those. They need to be linked into the
application and thereby have their module registration function hooked
up in order to be reachable at all.

Task-number: QTBUG-140220
Change-Id: I82e44fba2f2880ada16a5ea60e2e33f1900b208e
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit d66cfa93eefa185aa15d42e9dc38cbcdbd0578cb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>