qmllint: Consider "on-bindings" for required properties
The runtime doesn't complain if a required property is satisfied via a
ValueSource or an Interceptor, so qmllint shouldn't complain either.

Whether this really should apply to interceptors is debatable, but for
now let's aim for consistency. And that a ValueSource can satifsy the
requirement is arguably to be expected.

Pick-to: 6.8
Fixes: QTBUG-139632
Change-Id: I355b8a48e6b9bc2749839267c3b799e8eba5d72f
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
(cherry picked from commit 46925f589fadf9f16c4db2462e086f28a104f629)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>