Fix IME selection cursor update
The fix for QTBUG-94253 caused a regression in IME selection handling,
where the selection would not move the cursor when a commit string was
present. Cursor movement in this scenario is expected - for example, in
the word re-selection feature of the Qt Virtual Keyboard. The user
clicks inside the pre-edit text, and the VKB reselects a new segment up
to the clicked position. This is implemented using an IME event with
the commit string of the current pre-edit and the selection start
attribute set to the clicked position.

This patch refines the condition so that the original behavior of the
selection start parameter is handled correctly when the IME event
contains a commit string, with the exception of cases where an input
mask is present in the text input. This preserves the input mask
behavior fix from QTBUG-94253 while restoring correct selection
handling.

Pick-to: 6.8 6.5
Fixes: QTBUG-140400
Task-number: QTBUG-94253
Change-Id: I73d8ae32a1132d52cac9f39509ab2cbe922fa3d0
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Frederic Lefebvre <frederic.lefebvre@qt.io>
(cherry picked from commit 9324f445c0e94ef55906b84004962226a131968a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>