FluentWinUI3: Fix ComboBox size when the model changes
The problem was that enter Transition dynamically adjusts from and to
values and these values are not up-to-date. Also, transition animates
height property while from and to also depend on it.

Solve this by animating instead __heightScale with non changing from &
to values, so synchronizing property changes are not needed.

Do the same change for Menu that uses a similar approach.

Pick-to: 6.8
Task-number: QTBUG-138744
Change-Id: Ibede68a2d20ef437e0bc1637b5177f10c62935dd
Reviewed-by: Doris Verria <doris.verria@qt.io>
(cherry picked from commit 385e95794a14533eb580f1892d372a7bdd07ba89)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>