Fusion Button & RoundButton: bind icon.color to a palette color
The icon's color would adapt poorly to theme changes.
A good example is the QML media player in qtmultimedia, with has a few
buttons with icons and no text. When changing between a light and dark
theme, the icon would still have the same color, and there would barely
be any contrast between the icon and the button's background when using
a dark theme.

The binding created in this patch is taken from the Basic style.

Pick-to: 6.8
Change-Id: I922823a9109fd252c5a4ed84aba1c0a36adec263
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 68f2092b13f8956e77c830da4c3fa5e800b15624)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>