QQuickMacStyle: tweak NSTextFieldCell's drawing
This cell is drawing a frame/outline around a text field (a line edit in Qt).
Starting from macOS Tahoe this frame is slightly offset and with
larger size than requested (probably because now text field has
rounded corners and also the outline is thin and barely visible),
and as a result clipped by the current context's clip. In the
'Light' theme QLineEdit is completely invisible (unless it has a
focus). To address this issue we adjust the rect we draw line edit
with so it's the same as prior to Tahoe.

Note: this patch is a cherry-pick of fb61543e6e2 in qtbase, which
fixes the exact same problem for Widgets.

Pick-to: 6.9 6.8 6.5
Fixes: QTBUG-138465
Change-Id: I5955f4304e9fcd56de993267ff8963f6b133f14a
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit 01f096cea97f1793e005311233b9555b24a8aba0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>