qmltc: Fix context assignment
If an object is not the document root we need to set the document root's
context as its outer context, so that IDs end up in the right place.
However, when calling a method on an object, we need to provide the
context the method was declared in, not the context the object was
instantiated in. Otherwise we get the wrong set of IDs.

Fixes: QTBUG-139142
Change-Id: Iaeba89c44c955ed09a9778e5d6cc4150e86793ba
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
(cherry picked from commit ceb49a8736bef5077c15b1154e14079e1b73e932)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>