VectorImage: Fix offset transformation of use nodes
Instead of just using transform property itself for this, the
<use> node in SVG has an additional offset property which is documented
to be a "final offset transformation".

We would previously apply this as the x and y coordinates of the item
in the generated Qt Quick code, which works when there is no other
transform on the item (or the only transform is another translation).

But since the x and y is applied before the other transforms,
if we scale, rotate, etc. the coordinate system, then we would get
the wrong position.

Change-Id: I72b05439c06619231cfc5d4ba5a639e06d7aa184
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit 22ac54787fb07b19e55a0a0f5d1af3c727817653)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>