Compiler: Actually read the value of a key passed to qmltypes
We always set isJavaScriptBuiltin to true if the key was present instead
of actually reading the value. This seems to have worked because the key
was only added when it was true and the default value is false.

Amends 578946437fb31edd6aa78612e9db8fed9a3eb210

Pick-to: 6.8
Change-Id: I91f533129d233b251e221bf05abdc919eb8957c6
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 72e12635d81cd86362a3eaf2951160550c3a3ab1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>