QQmlListWrapper: add createOwned
We are going to need a way to store a copy of a QQmlList in Binding,
which doesn't reflect changes to the list. For that, we'll use
QQmlListWrapper in its "owning" mode. Add a helper function to make its
construction easier.

Pick-to: 6.8
Task-number: QTBUG-140858
Change-Id: Iee8af6d5d8f9772366f44f2a29cf90dbeada8823
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 173ce96cc2e218a71982575d52adec09a5878de6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>