QQmlValueTypeWrapper: Mark as dirty if gadgetPtr goes away
In QQmlValueTypeWrapper::write, we clear the gadgetPtr if it didn't
exist before, and it was only created for write. That however leads to
an inconsistent isDirty state: The wrapper assumes it still has data
which is up-to-date, but that data is actually gone.

Avoid that issue by also marking the wrapper as dirty when the gadgetPtr
is reset in write.

Fixes: QTBUG-140414
Change-Id: Ie73b47ae49d4a26a43e2dbcc47365ae71a343109
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit c920059cbc0098b6218e814d445eac7da345b0e0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>