Remove qml-i18n example in qtdeclarative repo
The qml-i18n example and the qtdeclarative repo itself depend
on the LinguistTools CMake package, which is not available while
building qtdeclarative.

This results in:
- missing build coverage for the example
- a cyclic dependency, because qtdeclarative depends on qttools
  (because of LinguistTools), and qttools depends on qtdeclarative
  (because of qdoc needs)
- we try to find a module that will never be available, unless
  qtdeclarative is built together with qttools in a top-level build
- in certain cases, if the top-level build fails early enough, the
  cyclic dependency causes reconfigurations to fail

Remove the example, the lookup of LinguistTools, and the linguist doc
references to avoid all these issues.

Replace the qt_add_translations reference in the docs of the
QQmlApplicationEngine class with a snippet copy.

Pick-to: 6.8
Task-number: QTBUG-141465
Change-Id: Id88603a3753f53dfac9f350e981073effda15eb1
Reviewed-by: Masoud Jami <masoud.jami@qt.io>
(cherry picked from commit 5a1610444562ed90280ff666c4474e569d87ded8)
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>