RangeSlider: keep values integer if everything is integer
Based on 42687caf7bc7496b838995f3fa68194ca8323da3.

If a user uses integer values for stepSize, from and to, they
most likely want the actual values of the handles to be integers, too.
Detect this condition, and store it in a new boolean member. If the
condition is met, we round the value in valueAt() (which, due to
floating point math, might not be an exact integer).

Fixes: QTBUG-141090
Pick-to: 6.8
Change-Id: If9507534db8fceb113e7426b5d46c6a023919d66
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 89d269f4273b535beb06586eb9ba27aeb77faa2e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>