QQuickText: re-layout eliding text that grows in width
Laying out text is somewhat expensive, which is why the geometryChange
function tries to determine whether or not it's necessary.

However, when growing the width of multi-line eliding text, we would
over-optimize, and stop relayouting, when there would have been more
space for text.

This worked for single line text, since then the widthExceeded flag
would always be set to true, when the natural text width exceeds the
layouts width. However, we don't always set that flag for multi-line
text, thus it's possible that we skip the relayout step every time we
increase the width of the Item.

Fixes: QTBUG-140875
Pick-to: 6.8
Change-Id: I1f7758bed2c52044d0179a83d4bf4a774f22e4e3
Reviewed-by: Bror Wetlesen Vedeld <bror.vedeld@qt.io>
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
(cherry picked from commit 90ec978b12c912d4838804f6383f378195ed432f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>